Class and interface were moved to the deque package and deleted unit test without the body
